Refactor->move up common methods.
