+++ src/Comcast/PhpLegalLicenses/Command/DependencyLicenseCommand.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace Comcast\PhpLegalLicenses\Console;
+
+use RuntimeException;
+use Symfony\Component\Console\Command\Command;
+
+class DependencyLicenseCommand extends Command
+{
+    /**
+     * Verify that the composer.lock file exists.
+     *
+     * @return void
+     */
+    protected function verifyComposerLockFilePresent()
+    {
+        if (is_file(getcwd().'/composer.lock')) {
+            return;
+        }
+        
+        throw new RuntimeException('Composer Lock file missing! Please run composer install and try again.');
+    }
+    
+    /**
+     * Parses the composer.lock file to retrieve all installed packages.
+     *
+     * @return array
+     */
+    protected function parseComposerLockFile()
+    {
+        $path = getcwd().'/composer.lock';
+        $contents = file_get_contents($path);
+        
+        return json_decode($contents, true);
+    }
+    
+}
